Make clippy an optional dependency, fixing build issues
Clippy depends on the latest nightly builds, and can fail to compile
even with a combination of latest nightly with older clippy. This
change removes the dependency while retaining the ability to use
clippy during development. It also has the effect of allowing xi-tui
to compile with stable rust, whereas it was implicitely tied to
nightly before.

As a side-effect, all dependencies not frozen to a specific version
were updated to the latest as of the time of this commit.