Defines the clap app with options for file and core
Core option allows the user to specify the path to a core binary. It
defaults to a binary named "xi-core" which is assumed to be in your
PATH.

File is a required argument. It allows the user to name the file to be
edited by `xi`, rather than always using "test-file."