Adds support for all current backend requests in the recv thread
For now, this prevents us from crashing _immediately_ on set_style or
scroll_to. We still crash in the main loop when we try to parse these as
Update structs, but it's an improvement.

In addition, we now need to be able to tell what kind of request we are
parsing in the main loop, so for now we will send a serde_json Array
through the channel, where the first element is always the request
method, and the second element is the params for that request.