refactor rendering & handle cursor properly
1. Introduce a "dirty" state for the view. A view is "dirty" if it has
been updated but not rendered yet. Two things can be dirty: the line
cache, and the cursor. Making this distinction allows us to avoid
re-drawing the whole screen just to update the cursor position.

2. Move the rendering from screen.rs to view.rs and line.rs. Due to the
previous change, after drawing the view, we need to mark it as "clean".
That means we need to mutate the view, which would change force us to
pass a mutable reference of the view to the screen, for just updating
the state. Instead, it seems cleaner to pass stdout from the screen to
the view, and let it render itself and update its state accordingly.

3. Handle the cursor properly for ascii characters. Before, we were
printing "fake" cursors because we were not able to find where the
cursor should be set when a line contained tabs (\t). The problem is
that tabs have inconsistent length. We now handle this correctly.
However we assume tabs are 4 spaces large. So when using xi-tui, one
must run `tabs -4`.