get rid of so-called "fake cursors"
These fake cursors were initially a workaround before we had proper
cursor handling. I kept them around in prevision of multi-cursor support
in xi-core. But they will make style handling complicated because they
mess with the utf-8 codepoint indices in the lines, so let's get rid of
them and revisit multi-cursor when it actually lands. We'll probably
have to handle that the same way we currently handle the "real" cursor.