Use the insert_tab() and insert_newline() xrl commands
Currently, we require the user to configure their terminal tab width to
four columns wide. One way we can get around this is to expand tabs into
spaces. Xi-core can do this for us -- there's actually a
'translate_tabs_to_spaces' config option, which is enabled by default.
In order for it to work, though, we have to indicate to xi-core when we
receive a tab via the "insert_tab" BufferEvent, so this changeset does
so.

For consistency, we also send "insert_newline" on newlines.