Fix tabs in xi term (mostly) (#85)
* Use the insert_tab() and insert_newline() xrl commands

Currently, we require the user to configure their terminal tab width to
four columns wide. One way we can get around this is to expand tabs into
spaces. Xi-core can do this for us -- there's actually a
'translate_tabs_to_spaces' config option, which is enabled by default.
In order for it to work, though, we have to indicate to xi-core when we
receive a tab via the "insert_tab" BufferEvent, so this changeset does
so.

For consistency, we also send "insert_newline" on newlines.

* Mostly fix tab/cursor position issues

If the terminal tab settings don't match our concept of tab locations,
funny things can happen when we try to position the cursor on a line
with tabs.

We can work around this a couple of ways:

(1) We can expand tabs into spaces. Xi-core does this by default for
most file types (Makefiles are a notable exception). Spaces obviously
aren't affected by the terminal tabstop settings, so we avoid the
problem entirely.

(2) Alternatively, we can allow the user to specify the tab size. Rather
than assuming tabs are 4 columns wide, we'll assume that the user's
preferences.xiconfig and terminal settings match up. Views will now
store a tab_size setting from the last ConfigChanged event they
received, and this setting will be used to position the cursor.

Note that we'll still have problems if the terminal and xi-term tab
settings don't agree, but hopefully this will reduce their occurrence
somewhat.